♻ Refactor: Use ctx.go order, shorthand GetHeaders(), and move Ctx structs
